+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Bug fixes:
 - Fix mismatch between geometries and colors in ``plot()`` if an empty or missing
   geometry is present (#2224)
 - Escape special characters to avoid TemplateSyntaxError in ``explore()`` (#2657)
+- Fix `to_parquet`/`to_feather` to not write an invalid bbox (with NaNs) in the
+  metadata in case of an empty GeoDataFrame (#2653)
 
 Notes on (optional) dependencies: